api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-ninja>=1.4.3",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "uvicorn>=0.35.0",
    "whitenoise>=6.9.0",
    "drf_stripe_subscription>=1.2.2",
    "supabase>=2.0.0",
    "PyJWT>=2.8.0",
    "drf-spectacular>=0.27.0",
    "django-cors-headers>=4.3.0",
    "requests>=2.31.0",
]
